fix(api, sp): unify client role logic and error handling across layers
Standardized error-handling and validation flow in both SQL procedures and API
services to ensure consistent behavior for client and admin role operations.

Changes include:
- Refactored SQL stored procedures to use TRY/CATCH with THROW for structured
  exception propagation and transaction safety.
- Normalized error codes and messages for role-related validation.
- Unified API service layer (ClientRoleService, UserRoleHandler) to interpret
  SQL errors and map them to accurate HTTP 400/409 responses.
- Improved logging and response formatting to provide clearer debugging context.
- Added validation to prevent invalid role assignments and to ensure default
  role fallback consistency.
- Updated population script to guarantee presence of default “Client” role.

These fixes eliminate previous mismatches in client role behavior, improve
debug visibility, and ensure API and DB remain in sync when handling role
logic errors.

Refs #248.

+++ 02NFL-Fantasy_API/NFL-Fantasy_API/Models/DTOs/Fantasy/RosterDTOs.cs
@@ -23,7 +23,7 @@ namespace NFL_Fantasy_API.Models.DTOs.Fantasy
     /// </summary>
     public class AddPlayerToRosterResponseDTO
     {
-        public int RosterID { get; set; }
+        public long RosterID { get; set; }
         public string Message { get; set; } = "Jugador agregado al roster exitosamente.";
     }
 }
